+++ source/defaults_linux.go
@@ -22,7 +22,7 @@ const OS_CONFIG_PATH = SOUS_DIR + "/config_linux.toml"
 
 const config_file = `# the version to use by default when creating
 # a new order
-default_target = "4.0"
+default_target = "4.2"
 
 # these are example versions that may not
 # match your system. please add, remove or
@@ -33,7 +33,7 @@ path = "~/software/blender_3.0/blender"
 
 [[target]]
 name = "4.0"
-path = "~/software/blender_4.0/blender"
+path = "~/software/blender_4.2/blender"
 
 [[target]]
 name = "canary"