+++ changelog.md
@@ -1,5 +1,23 @@
 # Changelog
 
+## 0.2.0
+
+Full Release
+
+### Features
+
+- Improved then quality of Sous Chef's output: it's now much nicer to look at and provides more information.
+- Added configurable placeholder behaviour.
+
+### Bugs
+
+- Overwriting now matches the new placeholder behaviour and correctly defers to the file instead of being always forced in a binary fashion.
+- Percentage of output resolution is now accounted for and correctly reported and applied to output.
+
+### Internal
+
+- Removed `targets` sub-command to add new targets from the CLI. You still need to edit the `.souschef/config.toml` yourself anyway, so let's just commit to it.
+
 ## 0.1.2
 
 Patch Release